Unit tests moved from src folder
