+++ test/additional-data.spec.ts
@@ -1,5 +1,5 @@
 import {TestBed, ComponentFixture} from "@angular/core/testing";
-import {AdditionalDataComponent} from "../app/components/additional-data";
+import {AdditionalDataComponent} from "../src/app/components/additional-data";
 
 describe('Component: AdditionalDataComponent', () => {
    let fixture: ComponentFixture<AdditionalDataComponent>, component, element, de, context;
